........
  r14716 | girazu | 2011-01-28 16:41:18 +0100 (vie, 28 ene 2011) | 1 line
  
  Añadidos iconos varios
........
  r14717 | girazu | 2011-01-28 16:58:19 +0100 (vie, 28 ene 2011) | 1 line
  
  Cambios en el menu de SMB
........
  r14718 | girazu | 2011-01-28 17:08:28 +0100 (vie, 28 ene 2011) | 1 line
  
  Cambios en el menu de CONSULTANT
........
  r14719 | eagirrezabal | 2011-01-31 09:52:59 +0100 (lun, 31 ene 2011) | 1 line
  
  el id al que hacia referencia el datascroller era erroneo.
........
  r14721 | ecastellano | 2011-01-31 11:11:01 +0100 (lun, 31 ene 2011) | 2 lines
  
  RESOLVED - bug 2189: Día de pago fuera de rango en previsión de tesorería 
  http://bugzilla.esferalia.com/show_bug.cgi?id=2189
........
  r14722 | eagirrezabal | 2011-01-31 11:16:42 +0100 (lun, 31 ene 2011) | 1 line
  
  Bug 2142 - Mejoras visuales en el listado a excel de seguimiento comercial
........
  r14727 | eagirrezabal | 2011-01-31 18:33:55 +0100 (lun, 31 ene 2011) | 1 line
  
  se modifica el metodo getDistribucionJornadas para que no unifique periodos NO consecutivos
........
  r14732 | girazu | 2011-01-31 19:31:14 +0100 (lun, 31 ene 2011) | 1 line
  
  Añadidos iconos varios
........
  r14733 | girazu | 2011-01-31 19:53:29 +0100 (lun, 31 ene 2011) | 1 line
  
  Mensajes
........
  r14734 | girazu | 2011-01-31 19:53:43 +0100 (lun, 31 ene 2011) | 1 line
  
  Añadidos iconos varios
........
  r14735 | girazu | 2011-01-31 19:59:03 +0100 (lun, 31 ene 2011) | 1 line
  
  Cambios en el menu de FISCAL
........
  r14736 | girazu | 2011-01-31 20:15:10 +0100 (lun, 31 ene 2011) | 1 line
  
  Cambios en el menu de FISCAL
........
  r14737 | girazu | 2011-01-31 20:15:48 +0100 (lun, 31 ene 2011) | 1 line
  
  Cambios en el menu de ACCOUNT
........
  r14738 | girazu | 2011-01-31 20:16:22 +0100 (lun, 31 ene 2011) | 1 line
  
  Cambios en el menu de ACCOUNT
........
  r14739 | girazu | 2011-01-31 20:18:43 +0100 (lun, 31 ene 2011) | 1 line
........
  r14742 | rtrepiana | 2011-02-01 11:30:47 +0100 (mar, 01 feb 2011) | 1 line
  
  Modificar los roles y perfiles existentes
........
  r14743 | rtrepiana | 2011-02-01 11:55:41 +0100 (mar, 01 feb 2011) | 1 line
  
  Modificar los roles y perfiles existentes
........
  r14747 | girazu | 2011-02-01 17:17:28 +0100 (mar, 01 feb 2011) | 1 line
  
  Mensajes
........
  r14748 | girazu | 2011-02-01 17:18:00 +0100 (mar, 01 feb 2011) | 1 line
  
  Resetear backProcess en el onReset y en el onSelect.
........
  r14749 | girazu | 2011-02-01 17:39:28 +0100 (mar, 01 feb 2011) | 2 lines
  
  ASSIGNED - bug 1956: FACTURA - Navegar de facturas a vencimientos 
  http://bugzilla.esferalia.com/show_bug.cgi?id=1956
........
  r14750 | girazu | 2011-02-01 17:40:25 +0100 (mar, 01 feb 2011) | 2 lines
  
  ASSIGNED - bug 1956: FACTURA - Navegar de facturas a vencimientos 
  http://bugzilla.esferalia.com/show_bug.cgi?id=1956
........
  r14751 | atellitu | 2011-02-01 17:50:19 +0100 (mar, 01 feb 2011) | 2 lines
  
  RESOLVED - bug 2187: No se parsean bien las imagenes embebidas en los mensajes 
  http://bugzilla.esferalia.com/show_bug.cgi?id=2187
........

+++ aon-ui-webmail/src/main/java/com/code/aon/ui/webmail/controller/MessageController.java
@@ -75,7 +75,6 @@ import com.code.aon.webmail.bean.AonMessage;
 import com.code.aon.webmail.bean.AonMessageUtils;
 import com.code.aon.webmail.bean.AonServer;
 import com.code.aon.webmail.bean.BundleConstants;
-import com.code.aon.webmail.bean.IMimeType;
 import com.code.aon.webmail.dao.IWebMailAlias;
 import com.sun.mail.imap.AppendUID;
 import com.sun.mail.imap.IMAPFolder;
@@ -281,12 +280,14 @@ public class MessageController implements IWebMailConstants, BundleConstants {
 	private void finishMessage() {
     	setShowNewMessageWindow(false);
 		content = null;
-		for( AonFile af : newMsgFileList ) {
-			if ( af.getFile().exists() ) {
-				FileUtils.deleteQuietly( af.getFile() );
+		if ( newMsgFileList == null ) {
+			for( AonFile af : newMsgFileList ) {
+				if ( af.getFile().exists() ) {
+					FileUtils.deleteQuietly( af.getFile() );
+				}
 			}
+	    	newMsgFileList = null;
 		}
-    	newMsgFileList = null;		
 	}
 
 	private void copyAttachmentsToFileList( AonMessage message ) throws WebmailException {
@@ -395,37 +396,51 @@ public class MessageController implements IWebMailConstants, BundleConstants {
 	//*********** END ATTACH ****************************************
 	//***************************************************************
 	
-	
-    public void onSend(ActionEvent event) {
-    	try {
-	    	AonMessage aonMessage = compoundMessage();
-	    	AonFolder dest = null;
-	    	AonServer server = getWebMailController().getServer();
-	    	try {
-	    		server.sendMessage(aonMessage);
-		    	dest = server.getAonFolder(server.getSentFolderName());
-		    	if (parentMessage!=null){
-			    	parentMessage.getMessage().setFlag(Flag.ANSWERED, true);
-			    	parentMessage.getParent().getFolder().expunge();
-		    	}
-	    	} catch ( WebmailException e ) {
-	    		LOGGER.error( e.getMessage(), e);
-		    	dest = server.getAonFolder(server.getDraftFolderName());
-	    	}
-	    	Message[] messages = new Message[1];
-    		messages[0] = aonMessage.getMessage();
-    		messages[0].setFlag(Flag.SEEN, true);
-	    	dest.open(Folder.READ_WRITE);
-	    	Folder desfFolder = dest.getFolder();
+	private void storeMessage( AonServer server, AonMessage aonMessage, boolean draft ) {
+    	Message[] messages = new Message[1];
+		messages[0] = aonMessage.getMessage();
+		try {
+			messages[0].setFlag(Flag.SEEN, true);
+			String folderName = draft ? server.getDraftFolderName() : server.getSentFolderName();
+			AonFolder folder = server.getAonFolder(folderName);
+			folder.open(Folder.READ_WRITE);
+	    	Folder desfFolder = folder.getFolder();
 	    	desfFolder.appendMessages(messages);
 	    	desfFolder.expunge();
-	    	dest.close(false);
+	    	folder.close(false);		
+		} catch (MessagingException e) {
+			LOGGER.error( "Error setting contactName", e );
+			AonUtil.addErrorMessage(e.getMessage());
+		}
+	}
+	
+    public void onSend(ActionEvent event) {
+    	AonServer server = getWebMailController().getServer();
+    	AonMessage aonMessage = null;
+    	try {
+	    	aonMessage = compoundMessage();
+    		server.sendMessage(aonMessage);
+		} catch (Throwable th) {
+			AonUtil.addErrorMessage(th.getMessage());
+			if ( aonMessage != null ) {
+	    		storeMessage(server, aonMessage, true);
+				refreshDraftFolder();	    		
+			}
+			throw new AbortProcessingException(th);
+		}
+    	try {	
+    		storeMessage(server, aonMessage, false);
+			if (parentMessage!=null){
+		    	parentMessage.getMessage().setFlag(Flag.ANSWERED, true);
+		    	parentMessage.getParent().getFolder().expunge();
+	    	}
 	    	deleteDraftMessage();
-	    	refreshDraftFolder();
-	    	finishMessage();
+    		refreshDraftFolder();
 		} catch (Throwable th) {
 			AonUtil.addErrorMessage(th.getMessage());
 			throw new AbortProcessingException(th);
+		} finally {
+	    	finishMessage();
 		}
     }
